fix(server): persist mod invocation nodes in export nodes
On a fresh analysis the parent contract needs to store then
resetup modified invocation nodes. This is managed with
connection type rules.

A multi-file test has been added to cover this case, it involves a
timeout to allow the second indexing of a parent file to test the
presistence of backlinks to modifier invocation over re-analysis.