feat(telemetry): add Hardhat telemetry setting
Add a Hardhat setting for telemetry. This is in addition to the global
telemetry setting. Both are communication to the language server.
Telemetry will only be sent if both are enabled (globally turning off
telemetry will stop Hardhat from sending).

Repurpose the show information message on FTUE. The result of the users
choice updates the extension hardhat setting. We record in the extension
globalState that the opt-in message has been seen by the user.