feat(hover): add hover text
This implements the `onHover` functionality of the LSP.

It does this for identifiers in local functions, and will give type
and other information from local variables, parameters and state
variables.

It adds hover for functions invocations, events and errors as well.

Comments are not included in the display (they are not in the ast).

Relates to #62