update logo and title (#311)
* chore: update extension logo

Use the new logo for the extension in the extension metadata - this will
get reflected in the marketplace.

* docs: update way extension is referred to

Remove the out of date internal docs references to **Hardhat for
VSCode**. The extension is now generally Solidity focused which is
reflected in our name change to **Solidity**. As there are other
**Solidity** extensions we disambiguate in our marketplace facing
documentation with **Solidity by Nomic Foundation**. Outside of the
marketplace we use **Solidity for VS Code by Nomic Foundation** to give
context that it is the vscode extension that we are talking about.

* feat: switch to solidity naming

The extension is providing general solidity features and is now called
`Solidity`, but in many places we were calling the extension `Hardhat
for VSCode`, or using hardhat naming.

This commit, changes the name to solidity:

* rename the output channels to `solidity`
* update commands/tasks with some `solidity` renames (we kept some
  hardhat for hardhat tasks)
* rename the Settings section to `solidity`