fix: limit status check and expand connected check
I backed out of the status check on send. I think we should have this
but I don't want to introduce a change I can't fully predict (is there
anything that was assuming non-running?).

At the same time, if you have a requestId, you should really cleanup the
handlers if an assertion fails. For the moment I have opted to avoid
hitting those assertions in send by making sure that each `send` with a
requestId also checks for connection.